+++ src/test/java/cn/yuqi/mini/spring/beans/factory/BeanFactoryTest.java
@@ -1,6 +1,7 @@
 package cn.yuqi.mini.spring.beans.factory;
 
 import cn.yuqi.mini.spring.beans.factory.config.BeanDefinition;
+import cn.yuqi.mini.spring.beans.factory.support.CglibSubclassingInstantiationStrategy;
 import cn.yuqi.mini.spring.beans.factory.support.DefaultListableBeanFactory;
 import org.junit.jupiter.api.Test;
 
@@ -9,12 +10,35 @@ class BeanFactoryTest {
     @Test
     void testGetBean() throws Exception {
         DefaultListableBeanFactory beanFactory = new DefaultListableBeanFactory();
-        BeanDefinition beanDefinition = new BeanDefinition(HelloService.class);
 
-        beanFactory.registerBeanDefinition("helloService",beanDefinition);
 
-        HelloService helloService = (HelloService) beanFactory.getBean("helloService");
-        helloService.sayHello();
+        BeanDefinition beanDefinition1 = new BeanDefinition(HelloService.class);
+
+        beanFactory.registerBeanDefinition("无参反射的helloService",beanDefinition1);
+
+        HelloService helloService1 = (HelloService) beanFactory.getBean("无参反射的helloService");
+        helloService1.sayHello();
+
+        BeanDefinition beanDefinition2 = new BeanDefinition(HelloService.class, "有参反射的构造传入的参数");
+        beanFactory.registerBeanDefinition("有参反射的helloService2",beanDefinition2);
+
+        HelloService helloService2 = (HelloService) beanFactory.getBean("有参反射的helloService2");
+        helloService2.sayHello();
+
+        // 尝试测试 cglib：
+        beanFactory.setInstantiationStrategy(new CglibSubclassingInstantiationStrategy());
+        BeanDefinition beanDefinition3 = new BeanDefinition(HelloService.class);
+
+        beanFactory.registerBeanDefinition("无参CGLIB的helloService",beanDefinition3);
+
+        HelloService helloService3 = (HelloService) beanFactory.getBean("无参CGLIB的helloService");
+        helloService3.sayHello();
+
+        BeanDefinition beanDefinition4 = new BeanDefinition(HelloService.class, "有参CGLIB的构造传入的参数");
+        beanFactory.registerBeanDefinition("有参CGLIB的helloService2",beanDefinition4);
+
+        HelloService helloService4 = (HelloService) beanFactory.getBean("有参CGLIB的helloService2");
+        helloService4.sayHello();
     }
 
     //此时如果使用下面这个类内定义的类进行测试会报错,需要一个顶级类