+++ src/test/java/cn/yuqi/mini/spring/beans/factory/HelloService.java
@@ -5,7 +5,10 @@ package cn.yuqi.mini.spring.beans.factory;
  * @date 2020/11/22
  */
 public class HelloService {
-
+    public HelloService() {}
+    public HelloService(String name){
+        System.out.println("HelloService 类的有参构造+" + name);
+    }
 	public String sayHello() {
 		System.out.println("hello");
 		return "hello";